Fixes address in use error for multiple test suites
Server attempted to listen to the same port multiple times when running test suites. Express app is abstracted into another file and test suites uses the app without listening to port.
