+++ backend/src/routes/__test__/user.route.test.ts
@@ -1,7 +1,7 @@
 import httpStatus from 'http-status';
 import databaseOperations from '../../utils/test/db-handler';
 import { UserModel } from 'src/models/user.model';
-import app from '../../index';
+import app from '../../server';
 
 const supertest = require('supertest');
